+++ src/main/java/com/example/gatherplan/appointment/service/UserService.java
@@ -7,4 +7,11 @@ public interface UserService {
 
     void joinUser(CreateUserReqDto createUserReqDto);
 
+    boolean checkHost(String appointmentCode, Long userId);
+
+    boolean checkJoin(String appointmentCode, Long userId);
+
+    boolean checkName(String appointmentCode, String name);
+
+    boolean checkNickname(String appointmentCode, String nickname);
 }
